protobufExcludeOutputs
**Problem**
When compiling complex protobuf schema, you often run into a situation
where you need some external schema during the compilation,
but you don't necessarily want to include the resulting Java output
as your own output because there's already a precompiled or well-known
bytecode version of the class.

A good example of this is classes under com/google/protobuf.

**Solution**
As a workaround to this situation, this implements protobufExcludeOutputs,
which is a seq of glob expression.
As a demonstration this key is initialized to be:
`Glob((ProtobufConfig / javaSource).value) / "com" / "google" / "protobuf" / "*.java"`.
This can include com/google/protobuf into protoc but will exclude it
from during source generation.